Revert "Merge branch 'overworld-expansion' of https://github.com/ghoulslash/pokeemerald into maps3"
This reverts commit 674abc96c57944cc0a431b8d01b458c55d26f57d, reversing
changes made to e1184af2118203e7b0ae12e87958e8bcd82baa79.

+++ src/easy_chat.c
@@ -27,6 +27,7 @@
 #include "task.h"
 #include "text_window.h"
 #include "window.h"
+#include "constants/easy_chat.h"
 #include "constants/event_objects.h"
 #include "constants/lilycove_lady.h"
 #include "constants/mauville_old_man.h"
@@ -4988,7 +4989,7 @@ static void HideStartSelectButtons(void)
 
 static void TryAddInterviewObjectEvents(void)
 {
-    u16 graphicsId;
+    int graphicsId;
     u8 spriteId;
 
     switch (GetDisplayedPersonType())
